server/requirements.txt
Flask
beautifulsoup4
requests
lxml
Flask-CORS 
Pillow 